shared/moe_experts.flash: cast output back to hidden_states.dtype
Root cause of the DSV4 MoE flash dtype-leak that surfaced as
aclnnGroupedMatmul "x dtype FLOAT, weight dtype BF16" crash at layer 2.

The flash chain ends with:
    final_hidden_states = npu_moe_token_unpermute(
        output.to(routing_weights.dtype),
        row_ids_map, probs=routing_weights,
    )
    return final_hidden_states          # ← returned in routing_weights dtype

DSV4's TopKRouter computes scores in fp32 deliberately (router math
needs the precision; sqrtsoftplus + scaling are fp32). So top_k_weights
flowing into experts is fp32. The unpermute did the weighted-sum in
fp32 (good, matches HF eager's bf16 * fp32 -> fp32 promote), but
returned without casting back. Eager fixes this with
``final.index_add_(0, token_idx, current.to(final.dtype))``; this path
forgot the trailing ``.to(hidden_states.dtype)``.

Symptom under DSV4 4-layer test:

    Layer 0 mlp input:  bf16   (good)
    Layer 0 flash out:  fp32   ← LEAK
    Layer 0 mlp out:    fp32   (routed_fp32 + shared_experts_bf16 -> fp32)
    Layer 1 mlp input:  fp32   (residual stream now corrupted)
    Layer 1 flash gets fp32 hidden_states -> npu_grouped_matmul rejects

Found via tests/npu/diagnose_dsv4_dtype_chain (forward hooks + flash
entry monkey-patch). Verified by reading HF reference _Experts.forward
which has the equivalent ``.to(final.dtype)`` step.

Fix: one ``.to(hidden_states.dtype)`` at the end. Multiplication still
happens in routing_weights' dtype (precision preserved), result casts
back to input dtype (residual stream contract preserved).

Qwen3.5 MoE didn't hit this because its router returns weights in
bf16 (no .float() cast in the score function). DSV4 with fp32 router
is the first caller to exercise the mixed-dtype output path.

+++ hf_npu_binder/shared/moe_experts.py
@@ -121,13 +121,25 @@ def flash(
     output = gmm_flash(activated, down_weight, tokens_per_expert)
 
     # 6. Scatter tokens back to their original positions, applying the
-    #    routing weights as part of the unpermute.
+    #    routing weights as part of the unpermute. The ``output.to(...)`` /
+    #    ``routing_weights`` pair runs the weighted-sum in routing_weights'
+    #    dtype (typically fp32, since DSV4-style routers compute scores in
+    #    fp32 for numerical stability) — equivalent to HF eager's
+    #    ``F.linear(...) * top_k_weights`` mixed-dtype multiply. The
+    #    ``.to(hidden_states.dtype)`` at the end mirrors HF eager's
+    #    ``current.to(final.dtype)`` before index_add: it casts the final
+    #    sum back to the residual stream dtype so callers see input-dtype
+    #    in, input-dtype out (the contract every other experts impl in
+    #    transformers/integrations/moe.py also satisfies). Without this
+    #    cast, fp32 routing weights silently promote the entire residual
+    #    stream layer-by-layer, eventually surfacing as
+    #    aclnnGroupedMatmul x.dtype != weight.dtype crashes downstream.
     final_hidden_states = torch_npu.npu_moe_token_unpermute(
         output.to(routing_weights.dtype),
         row_ids_map,
         probs=routing_weights,
     )
-    return final_hidden_states
+    return final_hidden_states.to(hidden_states.dtype)
 
 
 __all__ = ["flash", "_expert_weight_in_out"]
